requirements.txt
xlsxwriter
pytesseract
Pillow
opencv-python
tqdm
numpy